bs1 ok

+++ app/library/session.php
@@ -34,29 +34,27 @@ class Session
         $now = date(DATE_ISO8601);
         $this->phalcon_session = new SessionAdapter(['uniqueId' => Session::SESSION_NAME]);
         $this->phalcon_session->start();
-        $now = date(DATE_ISO8601);
-        $this->phalcon_session->set('auth', true);
-        $this->phalcon_session->set('user_ip', $this->get_userip());
-        $this->phalcon_session->set('last_active', $now);
-        $this->phalcon_session->has('user_agent') || $this->phalcon_session->set('user_agent', get_browser( null,true)[comment]);
 
-    }
-
-    public function init()
-    {
-        $deff_time = strtotime('now') - strtotime($this->phalcon_session->get('last_active'));
-
-        if (!empty($this->session_config['lifetime']) && $deff_time > $this->session_config['lifetime']) {
-            $this->destroy();
-            return FALSE;
-        } else if ($deff_time > $this->session_config['interval_refresh']) {
-            $now = date(DATE_ISO8601);
-            $this->phalcon_session->set('user_ip', $this->get_userip());
-            $this->phalcon_session->set('last_active', $now);
+        if($this->has('last_active')){
+            $deff_time = strtotime('now') - strtotime($this->get('last_active'));
+            if (!empty($this->session_config['lifetime']) && $deff_time > $this->session_config['lifetime']) {
+                $this->destroy();
+                $this->set('auth', false);
+                return FALSE;
+            }
         }
-        return TRUE;
+        $now = date(DATE_ISO8601);
+        $this->has_set('key', sha1(rand(1, 999999)));
+        $this->has_set('auth', false);
+        $this->set('user_ip', $this->get_userip());
+        $this->set('last_active', $now);
+        $this->has('user_agent') || $this->set('user_agent', get_browser(null, true)[comment]);
+
+
+        return true;
 
     }
+
 
     public function has($name = FALSE)
     {
@@ -84,19 +82,22 @@ class Session
         return $item;
     }
 
+    public function has_set($name, $value)
+    {
+        if ($this->has($name) === FALSE) {
+            $this->set($name, $value);
+        }
+    }
+
 
     public function set($name, $value)
     {
-        if (!$this->init()) {
-            return FALSE;
-        }
-
         $this->phalcon_session->set($name, $value);
         return TRUE;
     }
 
 
-    public function muti_set(array $array_set)
+    public function multi_set(array $array_set)
     {
 
         if (is_array($array_set)) {
